Fix #19 #18
